Step 1: Created helper class DefaultGridBuilder Issue ref #115
