+++ src/main/java/core/GameDriver.java
@@ -21,7 +21,7 @@ public class GameDriver {
     private void startGame() {
         presenter.displayMessage("Game is starting...");
         // User input for grid size
-        Grid grid = Grid.defaultGrid(); // Temporary grid
+        Grid grid = DefaultGridBuilder.defaultGrid(); // Temporary grid
         while (!grid.allShipsAreSunk()) {
             presenter.displayGrid(grid);
             presenter.displayMessage("Insert a coordinate to shoot!");
